refactor backup and restore modals, separate items on settings page
