+++ package.json
@@ -72,7 +72,7 @@
     "bash-color": "0.0.4",
     "buffer-shims": "1.0.0",
     "class-transformer": "0.5.1",
-    "class-validator": "0.14.0",
+    "class-validator": "0.14.1",
     "commander": "11.1.0",
     "dayjs": "1.11.10",
     "fastify": "4.25.1",
@@ -100,7 +100,7 @@
     "@types/fs-extra": "^11.0.4",
     "@types/jest": "^29.5.11",
     "@types/lodash": "^4.14.202",
-    "@types/node": "^20.10.7",
+    "@types/node": "^20.11.0",
     "@types/node-schedule": "^2.1.5",
     "@types/passport-jwt": "^4.0.0",
     "@types/semver": "^7.5.6",
@@ -108,15 +108,15 @@
     "@types/tar": "^6.1.10",
     "@types/tcp-port-used": "^1.0.4",
     "@types/unzipper": "^0.10.9",
-    "@typescript-eslint/eslint-plugin": "^6.18.0",
-    "@typescript-eslint/parser": "^6.18.0",
+    "@typescript-eslint/eslint-plugin": "^6.18.1",
+    "@typescript-eslint/parser": "^6.18.1",
     "babel-jest": "^29.7.0",
     "concurrently": "^8.2.2",
     "eslint": "^8.56.0",
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-import-newlines": "^1.3.4",
-    "eslint-plugin-jest": "^27.6.1",
+    "eslint-plugin-jest": "^27.6.3",
     "eslint-plugin-sort-exports": "^0.8.0",
     "form-data": "^4.0.0",
     "jest": "^29.7.0",
