+++ ui/package.json
@@ -49,7 +49,7 @@
     "qrcode": "1.5.3",
     "rxjs": "7.8.1",
     "semver": "7.5.4",
-    "socket.io-client": "4.7.3",
+    "socket.io-client": "4.7.4",
     "tslib": "2.6.2",
     "uuid": "9.0.1",
     "xterm": "4.19.0",
@@ -71,7 +71,7 @@
     "@types/emoji-js": "^3.5.2",
     "@types/file-saver": "^2.0.7",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.10.7",
+    "@types/node": "^20.11.0",
     "@types/qrcode": "^1.5.5",
     "@types/semver": "^7.5.6",
     "@types/socket.io-client": "^3.0.0",
