--- ui/src/app/core/backup-restore/backup-restore.component.html
@@ -1,72 +1,0 @@
-<div class="modal-content hb-backup-modal">
-  <div class="modal-header">
-    <h5 class="modal-title" [translate]="'backup.title_backup_restore'"></h5>
-    <button type="button" class="close" data-dismiss="modal" [attr.aria-label]="'form.button_close' | translate"
-      (click)="activeModal.dismiss('Dismiss')" *ngIf="!restoreInProgress">
-      <span aria-hidden="true">&times;</span>
-    </button>
-  </div>
-  <div class="modal-body" *ngIf="!restoreStarted && !setupWizardRestore">
-    <h4 class="text-center" [translate]="'backup.title_backup'"></h4>
-    <p [translate]="'backup.message_backup_help_one'"></p>
-    <p [translate]="'backup.message_backup_help_two'"></p>
-    <p class="primary-text" [translate]="'backup.message_backup_warning'"></p>
-    <div class="text-center">
-      <button type="button" class="btn btn-primary" data-dismiss="modal" (click)="onDownloadBackupClick()"
-        [disabled]="clicked">
-        <i *ngIf="clicked" class="fa fa-fw fa-circle-notch fa-spin"></i>
-        <span *ngIf="!clicked" [translate]="'backup.button_download_backup_archive'"></span>
-      </button>
-    </div>
-    <hr>
-    <h4 class="text-center" [translate]="'backup.title_restore'"></h4>
-    <p [translate]="'backup.message_restore_help_one'"></p>
-    <p [translate]="'backup.message_restore_help_two'"></p>
-    <div class="custom-file">
-      <input type="file" class="custom-file-input" id="restoreFileUpload" accept="application/gzip, .gz, .hbfx"
-        (change)="handleRestoreFileInput($event.target.files)">
-      <label class="custom-file-label" for="restoreFileUpload">{{ selectedFile?.name ||
-        'backup.label_choose_backup_file_to_restore' | translate }}</label>
-    </div>
-    <p class="primary-text mt-3" *ngIf="selectedFile" [translate]="'backup.message_restore_warning'"></p>
-    <div class="text-center" *ngIf="selectedFile">
-      <button type="button" class="btn btn-primary" data-dismiss="modal" (click)="onRestoreBackupClick()"
-        [disabled]="clicked">
-        <span *ngIf="clicked">
-          <i *ngIf="!uploadPercent || uploadPercent === 100" class="fa fa-fw fa-circle-notch fa-spin"></i>
-          <span *ngIf="uploadPercent && uploadPercent !== 100">{{ uploadPercent }}% - </span>
-          <span *ngIf="uploadPercent === 100"> Extracting Archive</span>
-          <span *ngIf="uploadPercent !== 100"> {{ 'backup.label_uploading' | translate }}</span>
-        </span>
-        <span *ngIf="!clicked" [translate]="'backup.button_restore_backup'"></span>
-      </button>
-    </div>
-  </div>
-
-  <div [hidden]="!restoreStarted" id="plugin-log-output" class="modal-body"></div>
-
-  <div class="modal-footer justify-content-between" *ngIf="!restoreStarted || restoreFailed === true">
-    <div class="text-left">
-      <button type="button" class="btn btn-elegant" data-dismiss="modal" (click)="activeModal.dismiss('Dismiss')"
-        [attr.aria-label]="'form.button_close' | translate">
-        {{ 'form.button_close' | translate }}
-      </button>
-    </div>
-    <div class="text-center"></div>
-    <div class="text-right">
-      <button type="button" class="btn btn-primary mr-auto" (click)="openScheduledBackups()">
-        {{ 'backup.title_scheduled_backups' | translate }}
-      </button>
-    </div>
-  </div>
-
-  <div class="modal-footer justify-content-between" *ngIf="!restoreInProgress && restoreStarted">
-    <div class="text-left"></div>
-    <div class="text-center">
-      <button type="button" class="btn btn-primary" data-dismiss="modal" (click)="postBackupRestart()">
-        {{ 'menu.hbrestart.title' | translate }}
-      </button>
-    </div>
-    <div class="text-right"></div>
-  </div>
-</div>
