+++ ui/src/app/modules/settings/restore/restore.component.html
@@ -1,0 +1,59 @@
+<div class="modal-content hb-backup-modal">
+  <div class="modal-header">
+    <h5 class="modal-title" [translate]="'backup.title_restore'"></h5>
+    <button type="button" class="close" data-dismiss="modal" [attr.aria-label]="'form.button_close' | translate"
+      (click)="activeModal.dismiss('Dismiss')" *ngIf="!restoreInProgress">
+      <span aria-hidden="true">&times;</span>
+    </button>
+  </div>
+  <div class="modal-body" *ngIf="!restoreStarted && !setupWizardRestore">
+    <p [translate]="'backup.message_restore_help_one'"></p>
+    <p [translate]="'backup.message_restore_help_two'"></p>
+    <p [translate]="'backup.message_restore_warning'"></p>
+    <div class="custom-file">
+      <input type="file" class="custom-file-input" id="restoreFileUpload" accept="application/gzip, .gz, .hbfx"
+        (change)="handleRestoreFileInput($event.target.files)">
+      <label class="custom-file-label" for="restoreFileUpload">{{ selectedFile?.name ||
+        'backup.label_choose_backup_file_to_restore' | translate }}</label>
+    </div>
+  </div>
+
+  <div [hidden]="!restoreStarted" id="plugin-log-output" class="modal-body"></div>
+
+  <div class="modal-footer justify-content-between" *ngIf="!restoreStarted || restoreFailed === true">
+    <div class="text-left">
+      <button type="button" *ngIf="selectedFile" class="btn btn-elegant" data-dismiss="modal" (click)="activeModal.dismiss('Dismiss')"
+        [attr.aria-label]="'form.button_close' | translate">
+        {{ 'form.button_close' | translate }}
+      </button>
+    </div>
+    <div class="text-center">
+      <button type="button" *ngIf="!selectedFile" class="btn btn-elegant" data-dismiss="modal" (click)="activeModal.dismiss('Dismiss')"
+        [attr.aria-label]="'form.button_close' | translate">
+        {{ 'form.button_close' | translate }}
+      </button>
+    </div>
+    <div class="text-right">
+      <button type="button" *ngIf="selectedFile" class="btn btn-primary" data-dismiss="modal" (click)="onRestoreBackupClick()"
+        [disabled]="clicked">
+        <span *ngIf="clicked">
+          <i *ngIf="!uploadPercent || uploadPercent === 100" class="fa fa-fw fa-circle-notch fa-spin"></i>
+          <span *ngIf="uploadPercent && uploadPercent !== 100">{{ uploadPercent }}% - </span>
+          <span *ngIf="uploadPercent === 100"> Extracting Archive</span>
+          <span *ngIf="uploadPercent !== 100"> {{ 'backup.label_uploading' | translate }}</span>
+        </span>
+        <span *ngIf="!clicked" [translate]="'backup.button_restore_backup'"></span>
+      </button>
+    </div>
+  </div>
+
+  <div class="modal-footer justify-content-between" *ngIf="!restoreInProgress && restoreStarted">
+    <div class="text-left"></div>
+    <div class="text-center">
+      <button type="button" class="btn btn-primary" data-dismiss="modal" (click)="postBackupRestart()">
+        {{ 'menu.hbrestart.title' | translate }}
+      </button>
+    </div>
+    <div class="text-right"></div>
+  </div>
+</div>
