+++ ui/src/app/modules/settings/settings.module.ts
@@ -3,9 +3,11 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
+import { BackupComponent } from '@/app/modules/settings/backup/backup.component';
 import { RemoveAllCachedAccessoriesModalComponent } from '@/app/modules/settings/remove-all-cached-accessories-modal/remove-all-cached-accessories-modal.component'; // eslint-disable-line max-len
 import { RemoveSingleCachedAccessoryModalComponent } from '@/app/modules/settings/remove-single-cached-accessory-modal/remove-single-cached-accessory-modal.component'; // eslint-disable-line max-len
 import { ResetHomebridgeModalComponent } from '@/app/modules/settings/reset-homebridge-modal/reset-homebridge-modal.component';
+import { RestoreComponent } from '@/app/modules/settings/restore/restore.component';
 import { SelectNetworkInterfacesComponent } from '@/app/modules/settings/select-network-interfaces/select-network-interfaces.component';
 import { SettingsRoutingModule } from '@/app/modules/settings/settings-routing.module';
 import { SettingsComponent } from '@/app/modules/settings/settings.component';
@@ -19,6 +21,8 @@ import { UnpairAccessoryModalComponent } from '@/app/modules/settings/unpair-acc
     RemoveAllCachedAccessoriesModalComponent,
     RemoveSingleCachedAccessoryModalComponent,
     SelectNetworkInterfacesComponent,
+    RestoreComponent,
+    BackupComponent,
   ],
   imports: [
     CommonModule,
